explicitly convert -1 to size_t in so_reduce_se_label_impl.h

+++ libtensor/symmetry/inst/so_reduce_se_label_impl.h
@@ -39,7 +39,7 @@ symmetry_operation_impl< so_reduce<N, M, T>, se_label<NM, T> >::do_perform(
     size_t nrsteps = 0;
     for (register size_t i = 0, j = 0; i < k_order1; i++) {
         if (params.msk[i]) {
-            map[i] = -1;
+            map[i] = size_t(-1);
             rmap[i] = params.rseq[i] + k_order2;
             nrsteps = std::max(nrsteps, params.rseq[i]);
         }
